+++ src/jawl_voicecompanion/__init__.py
@@ -13,7 +13,7 @@ from .vision import OpenAICompatibleVisionClient, VisionLookService
 from .presence import ScreenDeltaWatcher
 from .tts import CozyVoiceHttpClient, TTSService, TTSUnavailable, TTSCancelled
 from .voicemem_client import VoiceMemProcessClient, VoiceMemUnavailable
-from .jawl_web import JawlWebAdapter, JawlWebUnavailable
+from .jawl_web import JawlWebAdapter, JawlWebChatAdapter, JawlWebUnavailable
 
 __all__ = [
     "AccessLevel",
@@ -39,5 +39,6 @@ __all__ = [
     "VoiceMemProcessClient",
     "VoiceMemUnavailable",
     "JawlWebAdapter",
+    "JawlWebChatAdapter",
     "JawlWebUnavailable",
 ]